fix ListedValueAdder: capitalization of 1st word in custom value
if one of 'custom_values_to_rename' is entirely lowercase,
it would not match a lowercase custom value in feature profile
because the first word of the custom value was capitalized
programmatically, which will lead to non-match with a list.

This mistake was not caught by tests because an output feature
profile was not created at all and hence not checked against
gold standard.

Steps required:
1. Remove capitalization of custom value, compare it in lowercase
to values in lowercase.

2. In test module, check for presence of all output
feature profiles in output dir.